debt - allow to track disposables in running app (#192871)
* debt - allow to track disposables in running app

* add snapshot support

* add precondition

* update

* input
